+++ reporting.cpp
@@ -5,7 +5,7 @@
 #include "declaration.h"
 
 using namespace std;
-bool checkDisplace(long &step) {
+bool checkDisplace(unsigned long long &step) {
 
     if (dist(np.lastPairPos,np.position) > _bondEL) getAvailRec(availRec, np); // get all available receptors
     getAvailLig(availLig, ligands); // get all available ligands
